Revised stat generation

+++ src/main/java/utility/PStatistics.java
@@ -43,7 +43,7 @@ public interface PStatistics {
 
     String getEyeColor();
 
-    int getHeight();
+    double getHeight();
 
     void setCreativity(int creativity);
